+++ design_system/src/main/java/com/cairosquad/design_system/component/BottomNavItem.kt
@@ -1,0 +1,9 @@
+package com.cairosquad.design_system.component
+
+import androidx.annotation.DrawableRes
+
+data class BottomNavItem(
+    @DrawableRes val unColoredIcon: Int,
+    @DrawableRes val coloredIcon: Int,
+    val label: String
+)
