+++ design_system/src/main/java/com/cairosquad/design_system/theme/MovioTheme.kt
@@ -18,6 +18,7 @@ fun MovioTheme(
     isDarkTheme: Boolean = isSystemInDarkTheme(),
     content: @Composable () -> Unit
 ) {
+    UpdateStatusBarIconsForTheme(isDarkTheme)
     val theme = if (isDarkTheme) darkThemeColors else lightThemeColors
     CompositionLocalProvider(
         LocalMovioColor provides theme,
