+++ design_system/src/main/res/drawable/home_bottom_nav_colored.xml
