+++ design_system/src/main/res/drawable/library_bottom_nav_colored.xml
