+++ design_system/src/main/res/drawable/more_bottom_nav.xml
