+++ design_system/src/main/res/drawable/search_bottom_nav_colored.xml
